@@ -14,7 +14,7 @@
     "@types/angular": "1.5.8",
     "@types/angular-animate": "1.5.6",
     "@types/angular-gettext": "2.1.31",
-    "@types/angular-hotkeys": "0.0.31",
+    "@types/angular-hotkeys": "0.0.32",
     "@types/angular-sanitize": "1.3.4",
     "@types/autosize": "3.0.4",
     "@types/chart.js": "0.0.12",
@@ -25,7 +25,7 @@
     "@types/jstimezonedetect": "1.0.0",
     "@types/lodash": "0.0.27",
     "@types/moment-timezone": "0.2.34",
-    "@types/node": "7.0.12",
+    "@types/node": "7.0.14",
     "@types/nw.gui": "0.0.20",
     "@types/oclazyload": "0.0.31",
     "@types/qs": "6.2.31",
@@ -45,7 +45,7 @@
     "chart.js": "1.0.2",
     "codemirror": "5.25.2",
     "core-js": "2.4.1",
-    "date-fns": "1.28.3",
+    "date-fns": "1.28.4",
     "decompress-zip": "0.3.0",
     "express": "4.15.2",
     "favico.js": "0.3.10",
@@ -70,14 +70,14 @@
     "ua-parser-js": "0.7.12",
     "v-tooltip": "1.1.6",
     "vee-validate": "2.0.0-beta.25",
-    "vue": "2.2.6",
+    "vue": "2.3.0",
     "vue-class-component": "5.0.1",
     "vue-gettext": "gamejolt/vue-gettext#hworld-lazyload",
     "vue-property-decorator": "4.0.0",
-    "vue-router": "2.5.1",
-    "vue-server-renderer": "2.2.5",
+    "vue-router": "2.5.2",
+    "vue-server-renderer": "2.3.0",
     "vue-shortkey": "2.1.0",
-    "vue-ssr-webpack-plugin": "1.0.2",
+    "vue-ssr-webpack-plugin": "3.0.0",
     "vue-touch": "2.0.0-beta.4",
     "vuex": "2.3.1",
     "vuex-class": "0.1.3"
