@@ -1,16 +1,28 @@
 import Vue from 'vue';
 import { Component } from 'vue-property-decorator';
-import { State} from 'vuex-class';
+import { State } from 'vuex-class';
 import * as View from '!view!./auth.html?style=./auth.styl';
 import './auth-content.styl';
 
-import { store, Actions } from '../../store/index';
+import { store } from '../../store/index';
 import { MediaItem } from '../../../lib/gj-lib-client/components/media-item/media-item-model';
 import { AppCoverImg } from '../../components/cover-img/cover-img';
 import { makeObservableService } from '../../../lib/gj-lib-client/utils/vue';
 import { Environment } from '../../../lib/gj-lib-client/components/environment/environment.service';
 import { Connection } from '../../../lib/gj-lib-client/components/connection/connection-service';
 import { AppTranslateLangSelector } from '../../../lib/gj-lib-client/components/translate/lang-selector/lang-selector';
+import { Auth } from '../../../lib/gj-lib-client/components/auth/auth.service';
+
+export function loggedUserBlock()
+{
+	// Redirect right away if they are logged in.
+	if ( store.state.app.user ) {
+		Auth.redirectDashboard();
+
+		// Never resolve.
+		return new Promise(() => {});
+	}
+}
 
 @View
 @Component({
@@ -20,7 +32,7 @@ import { AppTranslateLangSelector } from '../../../lib/gj-lib-client/components/
 	},
 	async beforeRouteEnter( _to, _from, next )
 	{
-		await store.dispatch( Actions.bootstrap );
+		await store.dispatch( 'bootstrap' );
 		next();
 	},
 })
