@@ -1,11 +1,8 @@
-import '../vendor/base';
+import { app, router } from './bootstrap';
 
-import { platformBrowserDynamic } from 'ng-metadata/platform-browser-dynamic';
-import { enableProdMode } from 'ng-metadata/core';
-import { AppModule } from './app.module';
+// store.replaceState(window.__INITIAL_STATE__)
 
-if ( GJ_BUILD_TYPE === 'production' ) {
-	enableProdMode();
-}
-
-platformBrowserDynamic().bootstrapModule( AppModule );
+router.onReady( () =>
+{
+	app.$mount( '#app' );
+} );
